xen/dt: expose dt_unflatten_device_tree for runtime
SCF will be configured using partial device tree so we need to parse a DomU's
partial device tree in XEN. For this we should keep unflatten functions in a
persistent code section, not init.
In order to avoid memory leakage during unflattening device trees in runtime
we have to free memory used for the process. So return to upper layers a raw
memory pointer to let them free it once it is not needed.

Signed-off-by: Andrii Anisov <andrii_anisov@epam.com>