gic-vgic:vgic: avoid excessive conversions
Avoid excessive conversions between `pending_irq` and irq number and
priority. This is a tiny but pure optimization.

Also it aligns a `gic_raise_guest_irq()` function interface to its
current implementation in the meaning that it is removed a clearly
ignored priority argument. Seeing a call to this function from a
GICv3 code with an `p->lpi_priority` passed, it is assumed that LPI
priority should be used for prioritization in an `lr_pending` queue.
But the `lr_pending` queue is only sorted by `p->priority`, so it is
supposed that `p->lpi_priority` should be casted to `p->priority`
by GICv3 code to achieve expected behavior. On other hand, for LPIs,
`p->lpi_priority` might be still honored in the `gic_raise_guest_irq()`
functions, because it is still in the `pending_irq`.
But that fix should be made aside this patch.

Signed-off-by: Andrii Anisov <andrii_anisov@epam.com>