set pending and active state for HW IRQs as well
GIC document ARM IHI 0048B.b (ID072613), says to not set active and pending
state for HW IRQ, but does not say why. Theoretically we could miss sending
IRQ to guest, if it occurred in between guest handled, but not yet ended an
interrupt.
This also practically gives a slight gain in an interrupt extensive glmark2
benchmark.

Signed-off-by: Andrii Anisov <andrii_anisov@epam.com>