CK:xen: keep runstate area mapped
THIS IS A PATCH BASED ON RPM's PATCH.
INPROGRESS TO GET PHYSICAL RUNSTATE MAPPED

This is the patch from Roger Pau Monne.
The patch does map runstate area once, so that there is no need to
map that area each time hypervisor updates it.
This patch is rather controversial, there are doubts if VM could
change its runstate mappings without hypervisor notification.
As long as we still have no decission, no code cleanup is made as
well. See details here [1].

[1] https://lists.xenproject.org/archives/html/xen-devel/2019-02/msg00416.html

Signed-off-by: Andrii Anisov <andrii_anisov@epam.com>